check and change to track2
