+++ Vault/Library/FileTokenCache.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved. 
 // Licensed under the MIT License. See License.txt in the project root for license information. 
 
-using Microsoft.Identity.Client;
 using System;
 using System.Collections.Generic;
 using System.IO;
