+++ Vault/Library/SecretClientEx.cs
@@ -3,7 +3,6 @@
 
 using Azure.Security.KeyVault.Certificates;
 using Azure.Security.KeyVault.Secrets;
-using Microsoft.Azure.KeyVault;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -21,27 +20,6 @@ namespace Microsoft.Vault.Library
         public readonly string VaultUri;
 
         public SecretClientEx(string vaultName)
-        {
-            Utils.GuardVaultName(vaultName);
-            VaultName = vaultName;
-            VaultUri = string.Format(Consts.AzureVaultUriFormat, VaultName);
-        }
-
-        private string ToIdentifier(string endpoint, string name, string version) => $"{VaultUri}/{endpoint}/{name}" + (string.IsNullOrEmpty(version) ? "" : $"/{version}");
-
-        public string ToSecretIdentifier(string secretName, string version = null) => ToIdentifier(Consts.SecretsEndpoint, secretName, version);
-
-        public string ToCertificateIdentifier(string certificateName, string version = null) => ToIdentifier(Consts.CertificatesEndpoint, certificateName, version);
-
-        public override string ToString() => VaultUri;
-    }
-
-    internal class CertificateClientEx : CertificateClient
-    {
-        public readonly string VaultName;
-        public readonly string VaultUri;
-
-        public CertificateClientEx(string vaultName)
         {
             Utils.GuardVaultName(vaultName);
             VaultName = vaultName;
